Refactor package of spi.fixture (#9831)
