+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/spi/fixture/required/NoImplRequiredSPIFixture.java
@@ -15,13 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.spi.fixture;
+package org.apache.shardingsphere.infra.spi.fixture.required;
 
-public final class RequiredSPIFixtureDefaultFalseImpl implements RequiredSPIFixture {
-    
-    @Override
-    public boolean isDefault() {
-        return false;
-    }
+import org.apache.shardingsphere.infra.spi.required.RequiredSPI;
+
+public interface NoImplRequiredSPIFixture extends RequiredSPI {
 }
