+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/spi/fixture/required/RequiredSPIImpl.java
@@ -15,13 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.spi.fixture;
+package org.apache.shardingsphere.infra.spi.fixture.required;
 
-public final class RequiredSPIFixtureDefaultTrueImpl implements RequiredSPIFixture {
+import org.apache.shardingsphere.infra.spi.required.RequiredSPI;
+
+public final class RequiredSPIImpl implements RequiredSPI {
     
     @Override
     public boolean isDefault() {
-        return true;
+        return false;
     }
 }
-
