+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/spi/fixture/typed/TypedSPIFixtureImpl.java
@@ -15,10 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.spi.fixture;
+package org.apache.shardingsphere.infra.spi.fixture.typed;
 
-import org.apache.shardingsphere.infra.spi.typed.TypedSPI;
+import lombok.Getter;
+import lombok.Setter;
 
-public interface NoImplTypedSPIFixture extends TypedSPI {
+import java.util.Properties;
+
+@Getter
+@Setter
+public final class TypedSPIFixtureImpl implements TypedSPIFixture {
+    
+    private Properties props = new Properties();
+    
+    @Override
+    public String getType() {
+        return "FIXTURE";
+    }
 }
-
