+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/spi/ordered/OrderedSPIRegistryTest.java
@@ -19,8 +19,8 @@ package org.apache.shardingsphere.infra.spi.ordered;
 
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
 import org.apache.shardingsphere.infra.spi.fixture.FixtureCustomInterfaceImpl;
-import org.apache.shardingsphere.infra.spi.fixture.OrderedSPIFixture;
-import org.apache.shardingsphere.infra.spi.fixture.OrderedSPIFixtureImpl;
+import org.apache.shardingsphere.infra.spi.fixture.ordered.OrderedSPIFixture;
+import org.apache.shardingsphere.infra.spi.fixture.ordered.OrderedSPIFixtureImpl;
 import org.apache.shardingsphere.infra.spi.ordered.cache.OrderedServicesCache;
 import org.junit.After;
 import org.junit.Test;
