+++ shardingsphere-infra/shardingsphere-infra-common/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.spi.fixture.ordered.OrderedSPIFixture
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.spi.fixture.TypedSPIFixtureImpl
+org.apache.shardingsphere.infra.spi.fixture.ordered.OrderedSPIFixtureImpl
