+++ shardingsphere-infra/shardingsphere-infra-common/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.spi.fixture.typed.TypedSPIFixture
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.spi.fixture.OrderedSPIFixtureImpl
+org.apache.shardingsphere.infra.spi.fixture.typed.TypedSPIFixtureImpl
