+++ shardingsphere-infra/shardingsphere-infra-common/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.spi.required.RequiredSPI
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.spi.fixture.RequiredSPIImpl
+org.apache.shardingsphere.infra.spi.fixture.required.RequiredSPIImpl
